seperate related content module
